+++ azkustodata/query/v2/frame_reader_test.go
@@ -47,9 +47,9 @@ func TestDecodeValidFrames(t *testing.T) {
 		properties, err := query.ToStructs[QueryProperties](propertiesTable.Rows)
 		require.NoError(t, err)
 
-		require.Equal(t, 0, propertiesTable.TableId)
-		require.Equal(t, "QueryProperties", propertiesTable.TableKind)
-		require.Equal(t, "@ExtendedProperties", propertiesTable.TableName)
+		require.Equal(t, 0, propertiesTable.Header.TableId)
+		require.Equal(t, "QueryProperties", propertiesTable.Header.TableKind)
+		require.Equal(t, "@ExtendedProperties", propertiesTable.Header.TableName)
 		require.Equal(t, 1, properties[0].TableId)
 		require.Equal(t, "Visualization", properties[0].Key)
 		require.Equal(t, map[string]interface{}{"Accumulate": false, "AnomalyColumns": interface{}(nil), "IsQuerySorted": false, "Kind": interface{}(nil), "Legend": interface{}(nil), "Series": interface{}(nil), "Title": interface{}(nil), "Visualization": interface{}(nil), "XAxis": interface{}(nil), "XColumn": interface{}(nil), "XTitle": interface{}(nil), "Xmax": interface{}(nil), "Xmin": interface{}(nil), "YAxis": interface{}(nil), "YColumns": interface{}(nil), "YSplit": interface{}(nil), "YTitle": interface{}(nil), "Ymax": "NaN", "Ymin": "NaN"}, properties[0].Value)
@@ -142,9 +142,9 @@ func TestDecodeValidFrames(t *testing.T) {
 		queryCompletionInformation, err := query.ToStructs[QueryCompletionInformation](queryCompletionInformationTable.Rows)
 		require.NoError(t, err)
 
-		require.Equal(t, "QueryCompletionInformation", queryCompletionInformationTable.TableKind)
-		require.Equal(t, "QueryCompletionInformation", queryCompletionInformationTable.TableName)
-		require.Equal(t, 2, queryCompletionInformationTable.TableId)
+		require.Equal(t, "QueryCompletionInformation", queryCompletionInformationTable.Header.TableKind)
+		require.Equal(t, "QueryCompletionInformation", queryCompletionInformationTable.Header.TableName)
+		require.Equal(t, 2, queryCompletionInformationTable.Header.TableId)
 		require.Equal(t, 2, len(queryCompletionInformationTable.Rows))
 		require.Equal(t, "2023-11-26T13:34:17.0731478Z", queryCompletionInformation[0].Timestamp.Format(time.RFC3339Nano))
 		require.Equal(t, "blab6", queryCompletionInformation[0].ClientRequestId)
